+++ examples/resources/meraki_appliance_firewall_multicast_forwarding/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_firewall_multicast_forwarding.example "<network_id>"